requirements.txt
flask
torch
torchvision
pillow 
scikit-image 
requests 